refactor(terminal): split prompt lifecycle
